requirements.txt
streamlit
google-generativeai
ddgs
wikipedia